test(#3815): split the key-scoped refusal cases into their own module (#1135)
